+++ backend/model/stock.model.js
