+++ backend/model/strategy.model.js
